requirements.txt
asgiref==3.4.1
attrs==21.4.0
autopep8==1.5.7
#backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
pycodestyle==2.7.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-decouple==3.4
pytz==2021.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
tzdata==2021.5
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.7.0
cloudinary==1.29.0
